Display seller's info on ad detail page

+++ src/pages/Profile.js
@@ -78,16 +78,29 @@ const Profile = () => {
     getAds()
   }, [img])
 
+  // const deletePhoto = async () => {
+  //   const confirm = window.confirm('Delete photo permanently?')
+  //   if (confirm) {
+  //     await deleteObject(ref(storage, user.photoPath))
+  //     await updateDoc(doc(db, 'users', auth.currentUser.uid), {
+  //       photoUrl: '',
+  //       photoPath: '',
+  //     })
+  //   }
+  // }
+
   const deletePhoto = async () => {
-    const confirm = window.confirm('Delete photo permanently?')
-    if (confirm) {
-      await deleteObject(ref(storage, user.photoPath))
-      await updateDoc(doc(db, 'users', auth.currentUser.uid), {
-        photoUrl: '',
-        photoPath: '',
-      })
+    if (auth?.currentUser && auth.currentUser?.uid === id) {
+      const confirm = window.confirm("Delete photo permanently?");
+      if (confirm) {
+        await deleteObject(ref(storage, user.photoPath));
+        await updateDoc(doc(db, "users", auth.currentUser.uid), {
+          photoUrl: "",
+          photoPath: "",
+        });
+      }
     }
-  }
+  };
 
   // if uder exists, display user info else return null
   return user ? (
